+++ src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".activity.SplashActivity"
+            android:name=".UI.activity.SplashActivity"
             android:theme="@style/Theme.Design.Light.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,30 +24,30 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <!--<activity-->
+        <!--android:name=".UI.activity.AnimationActivity"-->
+        <!--android:label="@string/app_name"-->
+        <!--android:theme="@style/Theme.Design.Light.NoActionBar" />-->
+        <!--<activity-->
+        <!--android:name=".UI.activity.MainActivity"-->
+        <!--android:theme="@style/AppTheme.NoActionBar" />-->
         <activity
-            android:name=".activity.AnimationActivity"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.Design.Light.NoActionBar" />
-        <activity
-            android:name=".activity.MainActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
-            android:name=".activity.TabActivity"
+            android:name=".UI.activity.TabActivity"
             android:theme="@style/AppTheme.NoActionBar" />
 
-        <activity android:name=".activity.BaseActivity" />
+        <activity android:name=".BaseActivity" />
         <!--主Tab类-->
-        <activity android:name=".activity.TodayActivity" />
-        <activity android:name=".activity.StarActivity" />
-        <activity android:name=".activity.WebViewActivity" />
-        <activity android:name=".activity.TalkActivity" />
+        <activity android:name=".UI.activity.TodayActivity" />
+        <activity android:name=".UI.activity.StarActivity" />
+        <activity android:name=".UI.activity.WebViewActivity" />
+        <activity android:name=".UI.activity.TalkActivity" />
 
-        <activity android:name=".activity.FavoriteActivity" />
-        <activity android:name=".activity.WebFavoriteActivity" />
-        <activity android:name=".activity.FavoriteDetailActivity" />
-        <activity android:name=".activity.WebFavoriteDetailActivity" />
-        <activity android:name=".activity.NoteActivity" />
-        <activity android:name=".activity.NoteDetailActivity" />
+        <activity android:name=".UI.activity.FavoriteActivity" />
+        <activity android:name=".UI.activity.WebFavoriteActivity" />
+        <activity android:name=".UI.activity.FavoriteDetailActivity" />
+        <activity android:name=".UI.activity.WebFavoriteDetailActivity" />
+        <activity android:name=".UI.activity.NoteActivity" />
+        <activity android:name=".UI.activity.NoteDetailActivity" />
         <!--<activity android:name=".activity.Note_Activity" />-->
 
         <meta-data
