+++ src/main/java/com/leday/ActivityManager.java
@@ -1,4 +1,4 @@
-package com.leday.Util;
+package com.leday;
 
 import android.app.Activity;
 import android.content.Context;
