+++ src/main/java/com/leday/UI/activity/FavoriteActivity.java
@@ -1,4 +1,4 @@
-package com.leday.activity;
+package com.leday.UI.activity;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
@@ -11,6 +11,7 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
+import com.leday.BaseActivity;
 import com.leday.R;
 import com.leday.Util.PreferenUtil;
 import com.leday.entity.Today;
