+++ src/main/java/com/leday/UI/activity/WebFavoriteDetailActivity.java
@@ -1,60 +1,52 @@
-package com.leday.activity;
+package com.leday.UI.activity;
 
-import android.content.ContentValues;
 import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
-import android.view.KeyEvent;
 import android.view.View;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.leday.BaseActivity;
 import com.leday.R;
-import com.leday.Util.PreferenUtil;
 
-public class WebViewActivity extends BaseActivity implements View.OnClickListener {
+public class WebFavoriteDetailActivity extends BaseActivity implements View.OnClickListener {
 
     private WebView mWebView;
 
-    private String local_title, local_url;
-
-    @Override
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (keyCode == KeyEvent.KEYCODE_BACK) {
-            if (mWebView.canGoBack()) {
-                mWebView.goBack();
-            } else {
-                this.finish();
-            }
-        }
-        return true;
-    }
+    private String local_url,local_title,local_id;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_webview);
+        setContentView(R.layout.activity_webviewdetail);
 
         initView();
         initEvent();
     }
 
     private void initView() {
-        Intent intent = getIntent();
-        local_url = intent.getStringExtra("localurl");
-        local_title = intent.getStringExtra("localtitle");
+        local_title = getIntent().getStringExtra("local_title");
+        local_url = getIntent().getStringExtra("local_url");
+        local_id = getIntent().getStringExtra("local_id");
 
-        TextView mLike = (TextView) findViewById(R.id.txt_webview_like);
-        mWebView = (WebView) findViewById(R.id.webview_activity);
-        TextView mTitle = (TextView) findViewById(R.id.txt_webview_title);
+        ImageView mBack = (ImageView) findViewById(R.id.img_webviewdetail_back);
+        TextView mTitle = (TextView) findViewById(R.id.txt_webviewdetail_title);
+        TextView mUrl = (TextView) findViewById(R.id.txt_webviewdetail_url);
+        TextView mUnLike = (TextView) findViewById(R.id.txt_webviewdetail_like);
+        mWebView = (WebView) findViewById(R.id.webviewdetail_activity);
+
         mTitle.setText(local_title);
+        mUrl.setText("长按复制文章地址->" + local_url);
 
-        mLike.setOnClickListener(this);
+        mUnLike.setOnClickListener(this);
+        mBack.setOnClickListener(this);
     }
 
     private void initEvent() {
@@ -82,26 +74,18 @@ public class WebViewActivity extends BaseActivity implements View.OnClickListene
         });
     }
 
-    public void close(View view) {
-        finish();
-    }
-
     @Override
     public void onClick(View view) {
         switch (view.getId()) {
-            case R.id.txt_webview_like:
-                //建一张表保存文章
+            case R.id.txt_webviewdetail_like:
+                Snackbar.make(view, "取消成功", Snackbar.LENGTH_SHORT).show();
+                String local_delete = "DELETE FROM wechattb WHERE _id = '" + local_id + "'";
                 SQLiteDatabase mDatabase = openOrCreateDatabase("leday.db", MODE_PRIVATE, null);
-                mDatabase.execSQL("create table if not exists wechattb(_id integer primary key autoincrement,title text not null,url text not null)");
-                ContentValues mValues = new ContentValues();
-                mValues.put("title", local_title);
-                mValues.put("url", local_url);
-                mDatabase.insert("wechattb", null, mValues);
-                mValues.clear();
+                mDatabase.execSQL(local_delete);
                 mDatabase.close();
-                Snackbar.make(view, "收藏成功 : " + local_title, Snackbar.LENGTH_SHORT).show();
-                //权宜之计，做个标识给FavoriteActivity用
-                PreferenUtil.put(WebViewActivity.this, "wechattb_is_exist", "actually_not");
+                break;
+            default:
+                finish();
                 break;
         }
     }
