+++ src/main/java/com/leday/UI/activity/WebViewActivity.java
@@ -1,51 +1,61 @@
-package com.leday.activity;
+package com.leday.UI.activity;
 
+import android.content.ContentValues;
 import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
+import android.view.KeyEvent;
 import android.view.View;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
-import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.leday.BaseActivity;
 import com.leday.R;
+import com.leday.Util.PreferenUtil;
 
-public class WebFavoriteDetailActivity extends BaseActivity implements View.OnClickListener {
+public class WebViewActivity extends BaseActivity implements View.OnClickListener {
 
     private WebView mWebView;
 
-    private String local_url,local_title,local_id;
+    private String local_title, local_url;
+
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        if (keyCode == KeyEvent.KEYCODE_BACK) {
+            if (mWebView.canGoBack()) {
+                mWebView.goBack();
+            } else {
+                this.finish();
+            }
+        }
+        return true;
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_webviewdetail);
+        setContentView(R.layout.activity_webview);
 
         initView();
         initEvent();
     }
 
     private void initView() {
-        local_title = getIntent().getStringExtra("local_title");
-        local_url = getIntent().getStringExtra("local_url");
-        local_id = getIntent().getStringExtra("local_id");
+        Intent intent = getIntent();
+        local_url = intent.getStringExtra("localurl");
+        local_title = intent.getStringExtra("localtitle");
 
-        ImageView mBack = (ImageView) findViewById(R.id.img_webviewdetail_back);
-        TextView mTitle = (TextView) findViewById(R.id.txt_webviewdetail_title);
-        TextView mUrl = (TextView) findViewById(R.id.txt_webviewdetail_url);
-        TextView mUnLike = (TextView) findViewById(R.id.txt_webviewdetail_like);
-        mWebView = (WebView) findViewById(R.id.webviewdetail_activity);
-
+        TextView mLike = (TextView) findViewById(R.id.txt_webview_like);
+        mWebView = (WebView) findViewById(R.id.webview_activity);
+        TextView mTitle = (TextView) findViewById(R.id.txt_webview_title);
         mTitle.setText(local_title);
-        mUrl.setText("长按复制文章地址->" + local_url);
 
-        mUnLike.setOnClickListener(this);
-        mBack.setOnClickListener(this);
+        mLike.setOnClickListener(this);
     }
 
     private void initEvent() {
@@ -73,18 +83,26 @@ public class WebFavoriteDetailActivity extends BaseActivity implements View.OnCl
         });
     }
 
+    public void close(View view) {
+        finish();
+    }
+
     @Override
     public void onClick(View view) {
         switch (view.getId()) {
-            case R.id.txt_webviewdetail_like:
-                Snackbar.make(view, "取消成功", Snackbar.LENGTH_SHORT).show();
-                String local_delete = "DELETE FROM wechattb WHERE _id = '" + local_id + "'";
+            case R.id.txt_webview_like:
+                //建一张表保存文章
                 SQLiteDatabase mDatabase = openOrCreateDatabase("leday.db", MODE_PRIVATE, null);
-                mDatabase.execSQL(local_delete);
+                mDatabase.execSQL("create table if not exists wechattb(_id integer primary key autoincrement,title text not null,url text not null)");
+                ContentValues mValues = new ContentValues();
+                mValues.put("title", local_title);
+                mValues.put("url", local_url);
+                mDatabase.insert("wechattb", null, mValues);
+                mValues.clear();
                 mDatabase.close();
-                break;
-            default:
-                finish();
+                Snackbar.make(view, "收藏成功 : " + local_title, Snackbar.LENGTH_SHORT).show();
+                //权宜之计，做个标识给FavoriteActivity用
+                PreferenUtil.put(WebViewActivity.this, "wechattb_is_exist", "actually_not");
                 break;
         }
     }
