+++ src/main/java/com/leday/UI/adapter/NoteAdapter.java
@@ -1,4 +1,4 @@
-package com.leday.adapter;
+package com.leday.UI.adapter;
 
 import android.content.Context;
 import android.content.Intent;
@@ -10,31 +10,31 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.leday.R;
-import com.leday.activity.NoteDetailActivity;
+import com.leday.UI.activity.NoteDetailActivity;
 import com.leday.entity.Note;
 
 import java.util.ArrayList;
 
-public class NoteRecyclerViewAdapter extends RecyclerView.Adapter<NoteRecyclerViewAdapter.MyViewHolder> {
+public class NoteAdapter extends RecyclerView.Adapter<NoteAdapter.MyViewHolder> {
 
     private ArrayList<Note> mList;
     private Context mContext;
     private LayoutInflater mLayoutInflater;
 
-    public NoteRecyclerViewAdapter(Context context, ArrayList<Note> list) {
+    public NoteAdapter(Context context, ArrayList<Note> list) {
         this.mList = list;
         this.mContext = context;
         this.mLayoutInflater = LayoutInflater.from(context);
     }
 
     @Override
-    public NoteRecyclerViewAdapter.MyViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public NoteAdapter.MyViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         View layout = mLayoutInflater.inflate(R.layout.item_note, parent, false);
         return new MyViewHolder(layout);
     }
 
     @Override
-    public void onBindViewHolder(NoteRecyclerViewAdapter.MyViewHolder holder, final int position) {
+    public void onBindViewHolder(NoteAdapter.MyViewHolder holder, final int position) {
         holder.mTime.setText(mList.get(position).getTime());
         holder.mContent.setText(mList.get(position).getTitle());
         holder.mItemView.setOnClickListener(new View.OnClickListener() {
