+++ src/main/res/layout/activity_animation_welcome.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.leday.activity.AnimationActivity">
+    tools:context=".UI.activity.AnimationActivity">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewpager_activity_animation"
