+++ src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activity.MainActivity">
+    tools:context=".UI.activity.MainActivity">
 
     <!--<android.support.design.widget.AppBarLayout-->
     <!--android:id="@+id/app_bar"-->
