+++ src/main/res/layout/activity_tab.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.leday.activity.TabActivity">
+    tools:context=".UI.activity.TabActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
