--- src/main/res/layout/animation_welcome_b.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ImageView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:src="@drawable/img_star_5" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab_animation"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal|bottom"
-        android:layout_marginBottom="64dp"
-        android:onClick="animationWelcome"
-        android:src="@android:drawable/ic_dialog_dialer"
-        app:backgroundTint="#3f51b5"
-        app:rippleColor="#ff8dff" />
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal|bottom"
-        android:layout_marginBottom="32dp"
-        android:gravity="center">
-
-        <ImageView
-            android:id="@+id/dot_welcome_a"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/dot_empty" />
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:layout_toRightOf="@id/dot_welcome_a"
-            android:src="@drawable/dot_full" />
-    </RelativeLayout>
-
-</FrameLayout>
